move flask code out
