+++ src/main/kotlin/com/vanillastar/vsbrewing/item/PotionFlaskItem.kt
@@ -1,29 +1,26 @@
 package com.vanillastar.vsbrewing.item
 
+import com.vanillastar.vsbrewing.block.FlaskBlock
+import com.vanillastar.vsbrewing.block.FlaskBlock.Companion.WATERLOGGED
+import com.vanillastar.vsbrewing.block.MOD_BLOCKS
 import com.vanillastar.vsbrewing.component.MOD_COMPONENTS
-import com.vanillastar.vsbrewing.potion.MILK_POTION_ID
-import com.vanillastar.vsbrewing.potion.potionContentsMatchId
 import java.util.Optional
-import kotlin.Float
-import kotlin.Pair
 import kotlin.String
-import kotlin.Unit
 import kotlin.jvm.optionals.getOrNull
 import kotlin.repeat
 import kotlin.streams.asStream
-import net.minecraft.SharedConstants
+import net.minecraft.advancement.criterion.Criteria
 import net.minecraft.block.Blocks
 import net.minecraft.component.DataComponentTypes
-import net.minecraft.component.type.AttributeModifiersComponent
 import net.minecraft.component.type.PotionContentsComponent
 import net.minecraft.entity.LivingEntity
-import net.minecraft.entity.attribute.EntityAttribute
-import net.minecraft.entity.attribute.EntityAttributeModifier
-import net.minecraft.entity.effect.StatusEffectInstance
-import net.minecraft.entity.effect.StatusEffectUtil
+import net.minecraft.entity.player.PlayerEntity
+import net.minecraft.fluid.Fluids
+import net.minecraft.item.AliasedBlockItem
 import net.minecraft.item.Item
 import net.minecraft.item.ItemGroup
 import net.minecraft.item.ItemGroups
+import net.minecraft.item.ItemPlacementContext
 import net.minecraft.item.ItemStack
 import net.minecraft.item.ItemUsage
 import net.minecraft.item.ItemUsageContext
@@ -33,16 +30,17 @@ import net.minecraft.particle.ParticleTypes
 import net.minecraft.potion.Potion
 import net.minecraft.potion.Potions
 import net.minecraft.registry.Registries
-import net.minecraft.registry.entry.RegistryEntry
 import net.minecraft.registry.tag.BlockTags
-import net.minecraft.screen.ScreenTexts
+import net.minecraft.server.network.ServerPlayerEntity
 import net.minecraft.server.world.ServerWorld
 import net.minecraft.sound.SoundCategory
 import net.minecraft.sound.SoundEvents
 import net.minecraft.stat.Stats
 import net.minecraft.text.Text
 import net.minecraft.util.ActionResult
-import net.minecraft.util.Formatting
+import net.minecraft.util.Hand
+import net.minecraft.util.TypedActionResult
+import net.minecraft.util.UseAction
 import net.minecraft.util.math.Direction
 import net.minecraft.world.World
 import net.minecraft.world.event.GameEvent
@@ -54,11 +52,11 @@ val POTION_FLASK_ITEM_METADATA_IN_CONTEXT: ModItemMetadataInContext = {
       /* previousItem= */ null,
       Registries.POTION.streamEntries()
           .flatMap { potion ->
-            (DrinkableFlaskItem.MAX_USES downTo DrinkableFlaskItem.MIN_USES)
+            (PotionFlaskItem.MAX_USES downTo PotionFlaskItem.MIN_USES)
                 .asSequence()
                 .map { remainingUses ->
                   ModItemGroupVisibilityMetadata(
-                      if (remainingUses < DrinkableFlaskItem.MAX_USES) {
+                      if (remainingUses < PotionFlaskItem.MAX_USES) {
                         ItemGroup.StackVisibility.SEARCH_TAB_ONLY
                       } else {
                         ItemGroup.StackVisibility.PARENT_AND_SEARCH_TABS
@@ -87,159 +85,94 @@ val POTION_FLASK_ITEM_METADATA_IN_CONTEXT: ModItemMetadataInContext = {
   ) {
     it.maxCount(1)
         .component(DataComponentTypes.POTION_CONTENTS, PotionContentsComponent.DEFAULT)
-        .component(MOD_COMPONENTS.flaskRemainingUsesComponent, DrinkableFlaskItem.MAX_USES)
+        .component(MOD_COMPONENTS.flaskRemainingUsesComponent, PotionFlaskItem.MAX_USES)
         .recipeRemainder(this.glassFlaskItem)
   }
 }
 
 /** [Item] for a potion-filled flask. */
-class PotionFlaskItem(settings: Settings) : DrinkableFlaskItem(settings) {
+class PotionFlaskItem(settings: Settings) : AliasedBlockItem(MOD_BLOCKS.flaskBlock, settings) {
   companion object {
-    fun appendPotionFlaskDataTooltip(
-        stack: ItemStack,
-        context: TooltipContext,
-        tooltip: MutableList<Text>,
-    ) {
-      val potionContents = stack.get(DataComponentTypes.POTION_CONTENTS)
-      val effects = potionContents?.effects
-      val showEffectsTooltip =
-          effects != null && !potionContentsMatchId(potionContents, MILK_POTION_ID)
+    /** Minimum number of uses for a flask item. */
+    const val MIN_USES = 1
 
-      if (showEffectsTooltip) {
-        appendEffectsTooltip(effects, context.updateTickRate) { tooltip.add(it) }
-      }
-      appendRemainingUsesTooltip(stack, tooltip)
-      if (showEffectsTooltip) {
-        appendUsageTooltip(effects) { tooltip.add(it) }
-      }
-    }
+    /** Maximum number of uses for a flask item. */
+    const val MAX_USES = 3
 
-    /**
-     * Builds and appends the item tooltip displaying this potion flask's effects.
-     *
-     * This logic is mostly copied from [PotionContentsComponent.buildTooltip], since we wish to
-     * insert our own tooltip within the default potion tooltip. Injecting into the function with a
-     * mixin would be too fragile in this case.
-     */
-    private fun appendEffectsTooltip(
-        effects: Iterable<StatusEffectInstance>,
-        tickRate: Float,
-        textConsumer: (Text) -> Unit,
-    ) {
-      if (effects.none()) {
-        textConsumer(Text.translatable("effect.none").formatted(Formatting.GRAY))
-        return
-      }
-
-      for (effect in effects) {
-        var mutableText = Text.translatable(effect.translationKey)
-        if (effect.amplifier > 0) {
-          mutableText =
-              Text.translatable(
-                  "potion.withAmplifier",
-                  mutableText,
-                  Text.translatable("potion.potency." + effect.amplifier),
-              )
-        }
-        if (!effect.isDurationBelow(SharedConstants.TICKS_PER_SECOND)) {
-          mutableText =
-              Text.translatable(
-                  "potion.withDuration",
-                  mutableText,
-                  StatusEffectUtil.getDurationText(effect, /* multiplier= */ 1.0f, tickRate),
-              )
-        }
-        textConsumer(mutableText.formatted(effect.effectType.value().category.formatting))
-      }
-    }
-
-    /**
-     * Builds and appends the item tooltip displaying what happens when this potion flask is used.
-     *
-     * This logic is mostly copied from [PotionContentsComponent.buildTooltip], since we wish to
-     * insert our own tooltip within the default potion tooltip. Injecting into the function with a
-     * mixin would be too fragile in this case.
-     */
-    private fun appendUsageTooltip(
-        effects: Iterable<StatusEffectInstance>,
-        textConsumer: (Text) -> Unit,
-    ) {
-      val effectDataList =
-          mutableListOf<Pair<RegistryEntry<EntityAttribute>, EntityAttributeModifier>>()
-      for (statusEffectInstance in effects) {
-        statusEffectInstance.effectType.value().forEachAttributeModifier(
-            statusEffectInstance.amplifier
-        ) { attribute, modifier ->
-          effectDataList.add(Pair(attribute, modifier))
-        }
-      }
-
-      if (effectDataList.none { (_, modifier) -> modifier.value != 0.0 }) {
-        return
-      }
-
-      textConsumer(ScreenTexts.EMPTY)
-      textConsumer(Text.translatable("potion.whenDrank").formatted(Formatting.DARK_PURPLE))
-
-      for ((attribute, modifier) in effectDataList) {
-        var modifierDisplayValue =
-            modifier.value *
-                when (modifier.operation) {
-                  EntityAttributeModifier.Operation.ADD_MULTIPLIED_BASE -> 100
-                  EntityAttributeModifier.Operation.ADD_MULTIPLIED_TOTAL -> 100
-                  else -> 1
-                }
-
-        var modifierTranslationKeyPrefix: String
-        var formatting: Formatting
-        if (modifier.value >= 0.0) {
-          modifierTranslationKeyPrefix = "attribute.modifier.plus"
-          formatting = Formatting.BLUE
-        } else {
-          modifierDisplayValue *= -1
-          modifierTranslationKeyPrefix = "attribute.modifier.take"
-          formatting = Formatting.RED
-        }
-
-        textConsumer(
-            Text.translatable(
-                    "${modifierTranslationKeyPrefix}.${modifier.operation.id}",
-                    AttributeModifiersComponent.DECIMAL_FORMAT.format(modifierDisplayValue),
-                    Text.translatable(attribute.value().translationKey),
-                )
-                .formatted(formatting)
-        )
-      }
-    }
+    /** Duration for drinking a flask item, in ticks. */
+    const val MAX_USE_TIME = 32
   }
 
   override fun getDefaultStack(): ItemStack {
     val stack = super.getDefaultStack()
+    stack.set(MOD_COMPONENTS.flaskRemainingUsesComponent, MAX_USES)
     stack.set(DataComponentTypes.POTION_CONTENTS, PotionContentsComponent(Potions.WATER))
     return stack
   }
 
-  /** This is mostly copied from [PotionItem.finishUsing]. */
-  override fun onFinishUsing(stack: ItemStack, world: World, user: LivingEntity) {
-    stack
-        .getOrDefault(DataComponentTypes.POTION_CONTENTS, PotionContentsComponent.DEFAULT)
-        .forEachEffect {
-          if (it.effectType.value().isInstant) {
-            it.effectType
-                .value()
-                .applyInstantEffect(user, user, user, it.amplifier, /* proximity= */ 1.0)
-          } else {
-            user.addStatusEffect(it)
+  override fun getMaxUseTime(stack: ItemStack, user: LivingEntity) = MAX_USE_TIME
+
+  override fun getUseAction(stack: ItemStack) = UseAction.DRINK
+
+  override fun use(world: World, user: PlayerEntity, hand: Hand): TypedActionResult<ItemStack> =
+      ItemUsage.consumeHeldItem(world, user, hand)
+
+  /**
+   * This is mostly copied from [PotionItem.finishUsing], except we try to decrement the flask's
+   * remaining uses instead of immediately emptying the flask.
+   */
+  override fun finishUsing(stack: ItemStack, world: World, user: LivingEntity): ItemStack {
+    val player = user as? PlayerEntity
+    if (player is ServerPlayerEntity) {
+      Criteria.CONSUME_ITEM.trigger(player, stack)
+    }
+
+    if (!world.isClient) {
+      stack
+          .getOrDefault(DataComponentTypes.POTION_CONTENTS, PotionContentsComponent.DEFAULT)
+          .forEachEffect {
+            if (it.effectType.value().isInstant) {
+              it.effectType
+                  .value()
+                  .applyInstantEffect(user, user, user, it.amplifier, /* proximity= */ 1.0)
+            } else {
+              user.addStatusEffect(it)
+            }
           }
-        }
+    }
+
+    player?.incrementStat(Stats.USED.getOrCreateStat(this))
+
+    if (player?.isInCreativeMode != true) {
+      // If applicable, decrement remaining uses instead of discarding item.
+      val remainingUses = stack.getOrDefault(MOD_COMPONENTS.flaskRemainingUsesComponent, MAX_USES)
+      if (remainingUses > MIN_USES) {
+        stack.set(MOD_COMPONENTS.flaskRemainingUsesComponent, remainingUses - 1)
+      } else {
+        stack.decrementUnlessCreative(/* amount= */ 1, player)
+      }
+      if (stack.isEmpty) {
+        return ItemStack(MOD_ITEMS.glassFlaskItem)
+      }
+    }
+
+    user.emitGameEvent(GameEvent.DRINK)
+    return stack
   }
 
   /**
-   * This is copied mostly from [PotionItem.useOnBlock], except we try to decrement the potion
-   * flask's remaining uses instead of immediately emptying the flask.
+   * Uses the flask on a block.
+   *
+   * This tries to place the flask as a block first, before trying to apply the flask on a block as
+   * an item.
    */
-  override fun useOnBlockAsItem(context: ItemUsageContext): ActionResult {
+  override fun useOnBlock(context: ItemUsageContext): ActionResult {
     val player = context.player
+    if (player != null && player.isSneaking) {
+      // Place only if sneaking, so as not to interfere with drinking the potion.
+      return this.place(ItemPlacementContext(context))
+    }
+
     val world = context.world
     val blockPos = context.blockPos
     val stack = context.stack
@@ -305,6 +238,17 @@ class PotionFlaskItem(settings: Settings) : DrinkableFlaskItem(settings) {
     world.setBlockState(blockPos, Blocks.MUD.defaultState)
     return ActionResult.success(world.isClient)
   }
+
+  override fun getPlacementState(context: ItemPlacementContext) =
+      super.getPlacementState(context)
+          ?.with(
+              FlaskBlock.LEVEL,
+              context.stack.getOrDefault(MOD_COMPONENTS.flaskRemainingUsesComponent, 0),
+          )
+          ?.with(
+              WATERLOGGED,
+              context.world.getFluidState(context.blockPos).fluid.matchesType(Fluids.WATER),
+          )
 
   /** This is copied from [PotionItem.getTranslationKey]. */
   override fun getTranslationKey(stack: ItemStack): String =
